Reorg plugins to mimic janitor package setup
