--- docs/compute.rst
@@ -1,8 +1,0 @@
-Compute API Reference
-=====================
-
-.. currentmodule:: gordon_gcp
-
-.. automodule:: gordon_gcp.compute
-
-.. autoclass:: gordon_gcp.compute.GCEEnricher
