--- docs/dns.rst
@@ -1,8 +1,0 @@
-DNS API Reference
-=================
-
-.. currentmodule:: gordon_gcp
-
-.. automodule:: gordon_gcp.dns
-
-.. autoclass:: gordon_gcp.dns.GDNSPublisher
