--- docs/pubsub.rst
@@ -1,8 +1,0 @@
-PubSub API Reference
-====================
-
-.. currentmodule:: gordon_gcp
-
-.. automodule:: gordon_gcp.pubsub
-
-.. autoclass:: gordon_gcp.pubsub.GPSEventConsumer
