+++ requirements.txt
@@ -1,2 +1,3 @@
-gordon-dns==0.0.1.dev0
+# NOTE: temporary until we package & release gordon core
+# gordon-dns==0.0.1.dev0
 jsonschema==2.6.0
